+++ apps/next/CHANGELOG.md
@@ -1,5 +1,19 @@
 # nextjs-dapp
 
+## 1.0.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @public-assembly/dao-utils@0.0.21
+
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @public-assembly/dao-utils@0.0.20
+
 ## 1.0.14
 
 ### Patch Changes