+++ packages/dao-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @public-assembly/dao-utils
 
+## 0.0.21
+
+### Patch Changes
+
+- Fixes errors in previous version where certain contract addresses were hardcoded.
+
+## 0.0.20
+
+### Patch Changes
+
+- Removes the useNounsProtocol hook. Adds provider logic from builder for useContract instances. Updates both READMEs.
+
 ## 0.0.19
 
 ### Patch Changes